+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+### Added
+- Express.js middleware
+
 ## [2.0.0-rc.2] 2024-05-04
 
 ### Added